Move each proxies into a weak map
