+++ packages/ember-metal/lib/index.js
@@ -41,6 +41,11 @@ export {
 } from './property_set';
 export { objectAt } from './array';
 export {
+  eachProxyFor,
+  eachProxyArrayWillChange,
+  eachProxyArrayDidChange
+} from './each_proxy';
+export {
   addListener,
   hasListeners,
   on,
